requirements.txt
flask==2.0.3
flask-sqlalchemy==2.5.1
python-dotenv==0.19.2
python-dateutil==2.8.2
flask-wtf==1.0.0
flask-login==0.5.0
flask-migrate==3.1.0
python-telegram-bot==13.7
werkzeug==2.0.3
